Fixed a circular dependency issue between the 2 modules and moved some grenade code to the RuntimeBpExample module
