+++ Source/RuntimeBlueprints/RuntimeBlueprints.Build.cs
@@ -6,20 +6,26 @@ public class RuntimeBlueprints : ModuleRules
 {
 	public RuntimeBlueprints(ReadOnlyTargetRules Target) : base(Target)
 	{
-		PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
+        PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
         PrivatePCHHeaderFile = "RuntimeBpPCH.h";
 
-        PublicDependencyModuleNames.AddRange(new string[] { "Core", "CoreUObject", "Engine", "JSON", "JsonUtilities", "InputCore", "RuntimeBpExample" });
+        PublicDependencyModuleNames.AddRange(new string[] { "Core", "CoreUObject", "Engine", "JSON", "JsonUtilities", "InputCore" });
 
 
         PrivateDependencyModuleNames.AddRange(new string[] {  });
 
 		// Uncomment if you are using Slate UI
 		PrivateDependencyModuleNames.AddRange(new string[] { "Slate", "SlateCore" });
-		
-		// Uncomment if you are using online features
-		// PrivateDependencyModuleNames.Add("OnlineSubsystem");
 
-		// To include OnlineSubsystemSteam, add it to the plugins section in your uproject file with the Enabled attribute set to true
-	}
+        // Uncomment if you are using online features
+        // PrivateDependencyModuleNames.Add("OnlineSubsystem");
+
+        // To include OnlineSubsystemSteam, add it to the plugins section in your uproject file with the Enabled attribute set to true
+
+        //PublicIncludePaths.Add( "RuntimeBlueprints" );
+        PublicIncludePaths.AddRange(new string[] { "RuntimeBlueprints", "RuntimeBpExample" });
+
+
+        PrivateIncludePathModuleNames.AddRange(new string[] { "RuntimeBpExample" });
+    }
 }
