+++ Source/RuntimeBlueprints/RuntimeBpData.h
@@ -12,7 +12,7 @@
 #include "GameFramework/Character.h"
 #include "Components/LightComponent.h"
 #include "Particles/ParticleSystemComponent.h"
-#include "RuntimeBpGrenade.h"
+#include "RuntimeBpExample/RuntimeBpGrenade.h"
 
 #include "RuntimeBpData.generated.h"
 
@@ -101,6 +101,7 @@ struct FNodeVarArgs
 	void Reset()
 	{
 		VariableData.Reset();
+		StringData = "";
 	}
 
 	FNodeVarArgs()
@@ -569,6 +570,11 @@ struct FNodeVarArgs
 		return false;
 	}
 
+	~FNodeVarArgs()
+	{
+		Reset();
+	}
+
 };
 
 // Array of FNodeVarArgs nested in a struct, this is used to store array variables
@@ -581,6 +587,11 @@ struct FNodeVarArgsArray
 	TArray<FNodeVarArgs> Array;
 
 	FNodeVarArgsArray()
+	{
+		Array.Empty();
+	}
+
+	~FNodeVarArgsArray()
 	{
 		Array.Empty();
 	}
